requirements.txt
torch>=2.0
torchvision>=0.15
pyyaml>=6.0
tqdm>=4.66
